CCFT: Rename filter file with hpp extension
