+++ examples/custom_content_filter/c++03/ccf_subscriber.cxx
@@ -15,7 +15,7 @@
 #include <dds/dds.hpp>
 #include <rti/core/ListenerBinder.hpp>
 #include "ccf.hpp"
-#include "filter.cxx"
+#include "filter.hpp"
 
 using namespace dds::core;
 using namespace dds::core::status;
