+++ examples/custom_content_filter/c++03/filter.hpp
